renamed pid_report module to child_proc_mgmt
