+++ apps/backend/pits_n_giggles.py
@@ -36,7 +36,7 @@ from apps.backend.telemetry_layer import initTelemetryLayer
 from apps.backend.ui_intf_layer import TelemetryWebServer, initUiIntfLayer
 from lib.config import load_config_from_ini
 from lib.error_status import PngError
-from lib.pid_report import report_pid_from_child
+from lib.child_proc_mgmt import report_pid_from_child
 from lib.version import get_version
 
 # -------------------------------------- GLOBALS -----------------------------------------------------------------------
