+++ lib/child_proc_mgmt.py
